# contracts/openapi/argus-v1.yaml
openapi: 3.1.0
info:
  title: Argus v1
  version: 1.0.0
  description: |
    End-of-day portfolio monitoring SaaS for retail investors.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - sessionCookie: []
    csrfToken: []
paths:
...
  /admin/users:
    post:
      tags: [admin]
      summary: Search admin users
      description: |
        POST (not GET) so the only sensitive criterion — `email_contains` — travels
        in the request body, out of access logs / proxy caches / browser history.
        Non-sensitive filters stay in query params.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - { in: query, name: is_suspended, schema: { type: boolean } }
        - { in: query, name: is_deleted, schema: { type: boolean } }
        - { in: query, name: is_verified, schema: { type: boolean } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: per_page, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                email_contains:
                  type: string
                  minLength: 1
                  maxLength: 254
                  description: Case-insensitive substring match against email.
      responses:
        "200":
          description: Users collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeAdminUserCollection"
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [admin]
      summary: Get a single user (admin view)
      security:
        - sessionCookie: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeAdminUser"
          description: User
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/users/{id}/suspend:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [admin]
      summary: Suspend a user (idempotent)
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeAdminUserAction"
          description: Suspend result
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/users/{id}/unsuspend:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [admin]
      summary: Unsuspend a user (idempotent)
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeAdminUserAction"
          description: Unsuspend result
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/users/{id}/delete:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [admin]
      summary: Soft-delete a user (idempotent)
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeAdminUserAction"
          description: Soft-delete result
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: argus_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  responses:
...
    Forbidden:
      description: Owner-scope or admin-scope violation, or account suspended / unverified
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource not found or not owned by caller
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/PaginationMeta" }
        links: { $ref: "#/components/schemas/PaginationLinks" }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        total_pages: { type: integer }
    PaginationLinks:
      type: object
      properties:
        self: { type: string }
        next: { type: string, nullable: true }
        prev: { type: string, nullable: true }
        last: { type: string }
...
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        is_verified: { type: boolean }
        is_suspended: { type: boolean }
        is_deleted: { type: boolean }
        is_admin: { type: boolean }
        created_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
    EnvelopeAdminUser:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/AdminUser" }
    EnvelopeAdminUserCollection:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/AdminUser" }
    EnvelopeAdminUserAction:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                id: { type: string, format: uuid }
                is_suspended: { type: boolean }
                is_deleted: { type: boolean }
                deleted_at: { type: string, format: date-time, nullable: true }
...